feat: reducer sync: initial (#18)

+++ co-primitives/src/types/date.rs
@@ -1,1 +1,1 @@
-pub type Date = u64; // unix timestamp
+pub type Date = u128; // unix timestamp in milliseconds
